+++ src_lua_interface/GenericContainerLuaInterface.cc
@@ -61,7 +61,7 @@ namespace GenericContainerNamespace {
   static
   inline
   bool isInteger( real_type x )
-  { return isZero(x-floor(x)) ; }
+  { return isZero( x-static_cast<long>(floor(x)) ) ; }
 
   /*
   //   _                _           ____  ____
